requirements.txt
# Core dependencies
python-dotenv==1.0.0
plotly==5.17.0
dash==2.14.0
pandas==2.1.3
requests==2.31.0

# Data processing
numpy==1.24.3

# API integrations
spotipy==2.23.0
google-api-python-client==2.105.0
google-auth-oauthlib==1.1.0

# Visualization enhancements
plotly-express==0.4.1
dash-bootstrap-components==1.5.0

# Development and testing
pytest==7.4.3
black==23.11.0
flake8==6.1.0
